Add files via upload
SVD is a powerful tool that finds applications in various fields due to its ability to represent data in a reduced and meaningful form, facilitating efficient computation and better understanding of underlying patterns in the data. However, it is computationally expensive for large matrices, and approximate methods like Truncated SVD are often used in such cases to retain only the most significant singular values and vectors.